Tabs component are moved to storybook
