+++ package.json
@@ -28,7 +28,7 @@
     "clsx": "1.1.1",
     "context-api": "0.0.2",
     "date-fns": "2.28.0",
-    "next": "12.1.6",
+    "next": "^12.1.6",
     "next-images": "1.8.4",
     "nprogress": "0.2.0",
     "numeral": "2.0.6",
