+++ src/modules/common/components/Tabs/stories/Tabs.stories.tsx
@@ -1,0 +1,88 @@
+
+import Tabs from '@mui/material/Tabs';
+import Tab from '@mui/material/Tab';
+import { SyntheticEvent, useState } from 'react';
+import { Meta } from '@storybook/react';
+import { Box, Typography } from '@mui/material';
+
+interface TabPanelProps {
+    children?: React.ReactNode;
+    index: number;
+    value: number;
+}
+
+function TabPanel(props: TabPanelProps) {
+    const { children, value, index, ...other } = props;
+
+    return (
+        <div
+            role="tabpanel"
+            hidden={value !== index}
+            id={`simple-tabpanel-${index}`}
+            aria-labelledby={`simple-tab-${index}`}
+            {...other}
+        >
+            {value === index && (
+                <Box sx={{ p: 3 }}>
+                    <Typography>{children}</Typography>
+                </Box>
+            )}
+        </div>
+    );
+}
+
+function a11yProps(index: number) {
+    return {
+        id: `simple-tab-${index}`,
+        'aria-controls': `simple-tabpanel-${index}`
+    };
+}
+
+const meta: Meta<typeof Tabs> = {
+    title: 'Common/Tabs',
+    component: Tabs,
+    args: {
+        children: 'Tabs',
+        className: 'font-bold'
+    }
+};
+
+export default meta;
+
+
+export const Basic = () => {
+    const [value, setValue] = useState(0);
+
+    const handleChange = (_event: SyntheticEvent, newValue: number) => {
+        setValue(newValue);
+    };
+    return (
+        <Box sx={{ width: '100%' }}>
+            <Tabs
+                variant="scrollable"
+                scrollButtons="auto"
+                textColor="primary"
+                indicatorColor="primary"
+                value={value}
+                onChange={handleChange}
+                aria-label="basic tabs example"
+            >
+                <Tab label="Item One" {...a11yProps(0)} />
+                <Tab label="Item Two" {...a11yProps(1)} />
+                <Tab label="Item Three" {...a11yProps(2)} />
+            </Tabs>
+            <TabPanel value={value} index={0}>
+                Item One
+            </TabPanel>
+            <TabPanel value={value} index={1}>
+                Item Two
+            </TabPanel>
+            <TabPanel value={value} index={2}>
+                Item Three
+            </TabPanel>
+        </Box>
+
+    );
+}
+
+
